Improved the example a bit
To make it more server like I spin up a go handler which behaves like a server waiting for a connection. The client 'connects' and gets a server greeting and the client then initates a request. This is a bit more server/client like. I also discovered that the net.Pipe Read needs to be in a loop:  it may only receive a message in fragments from the write end of the pipe whifunctionch causes stalls otherwise so I created a ReadAll function.